試行錯誤

+++ pci/src/xhc/allocator/aligned_address.rs
@@ -1,13 +1,13 @@
 use crate::error::{AllocateReason, PciError, PciResult};
 
-pub struct AlignedAddress(usize);
+pub struct AlignedAddress(u64);
 
 impl AlignedAddress {
-    pub fn new_uncheck(addr: usize) -> Self {
+    pub fn new_uncheck(addr: u64) -> Self {
         Self(addr)
     }
 
-    pub fn new_with_check_align_64_bytes(addr: usize) -> PciResult<Self> {
+    pub fn new_with_check_align_64_bytes(addr: u64) -> PciResult<Self> {
         if is_align_64_bytes(addr) {
             Ok(Self::new_uncheck(addr))
         } else {
@@ -21,7 +21,7 @@ impl AlignedAddress {
 
     pub fn address(&self) -> PciResult<u64> {
         if is_align_64_bytes(self.0) {
-            Ok(self.0 as u64)
+            Ok(self.0)
         } else {
             Err(PciError::FailedAllocate(
                 AllocateReason::NotAlignedAddress {
@@ -32,6 +32,6 @@ impl AlignedAddress {
     }
 }
 
-fn is_align_64_bytes(value: usize) -> bool {
+fn is_align_64_bytes(value: u64) -> bool {
     (value % 64) == 0
 }
